Move Leaderboard in the Leaderboard directory
